Move TestConsole to the test package, hide type in TestConsole.make, add tut text
